requirements.txt
Flask==0.10.1
GitHub-Flask==2.0.1
GitPython==2.1.0
requests==2.6.2
